+++ @es-labs/node/auth/index.js
@@ -6,27 +6,33 @@ const jwt = require('jsonwebtoken')
 
 // const uuid = require('uuid/v4')
 // const qrcode = require('qrcode')
-const { COOKIE_HTTPONLY, COOKIE_SAMESITE, COOKIE_SECURE, COOKIE_MAXAGE, COOKIE_DOMAIN } = global.CONFIG
-const { USE_OTP, OTP_EXPIRY, CORS_OPTIONS, AUTH_USER_STORE, AUTH_USER_STORE_NAME } = global.CONFIG
-const { AUTH_REFRESH_URL, AUTH_USER_FIELD_LOGIN, AUTH_USER_FIELD_PASSWORD, AUTH_USER_FIELD_GAKEY, AUTH_USER_FIELD_ID_FOR_JWT, AUTH_USER_FIELDS_JWT_PAYLOAD = ''} = global.CONFIG
-const { JWT_ALG, JWT_SECRET, JWT_REFRESH_SECRET, JWT_EXPIRY, JWT_REFRESH_EXPIRY, JWT_REFRESH_STORE ='keyv', JWT_CERTS, JWT_REFRESH_CERTS } = global.CONFIG
+let COOKIE_HTTPONLY, COOKIE_SAMESITE, COOKIE_SECURE, COOKIE_MAXAGE, COOKIE_DOMAIN,
+  USE_OTP, OTP_EXPIRY, CORS_OPTIONS,
+  AUTH_REFRESH_URL, AUTH_USER_FIELD_LOGIN, AUTH_USER_FIELD_PASSWORD, AUTH_USER_FIELD_GAKEY, AUTH_USER_FIELD_ID_FOR_JWT, AUTH_USER_FIELDS_JWT_PAYLOAD,
+  JWT_ALG, JWT_SECRET, JWT_REFRESH_SECRET, JWT_EXPIRY, JWT_REFRESH_EXPIRY, JWT_CERTS, JWT_REFRESH_CERTS,
+  JWT_REFRESH_STORE, AUTH_USER_STORE, AUTH_USER_STORE_NAME, JWT_REFRESH_STORE_NAME,
+  setRefreshToken, getRefreshToken, revokeRefreshToken, setRefreshTokenStoreName,
+  findUser, updateUser, setAuthUserStoreName
 
-let mongo, ObjectID
-if (AUTH_USER_STORE === 'mongo') {
-  mongo = require('../services/db/mongodb').get()
-  ObjectID = mongo.ObjectID
+const setupAuth = (options = global.CONFIG) => {
+  ({
+    COOKIE_HTTPONLY, COOKIE_SAMESITE, COOKIE_SECURE, COOKIE_MAXAGE, COOKIE_DOMAIN,
+    USE_OTP, OTP_EXPIRY, CORS_OPTIONS,
+    AUTH_REFRESH_URL, AUTH_USER_FIELD_LOGIN, AUTH_USER_FIELD_PASSWORD, AUTH_USER_FIELD_GAKEY, AUTH_USER_FIELD_ID_FOR_JWT, AUTH_USER_FIELDS_JWT_PAYLOAD = '',
+    JWT_ALG, JWT_SECRET, JWT_REFRESH_SECRET, JWT_EXPIRY, JWT_REFRESH_EXPIRY, JWT_CERTS, JWT_REFRESH_CERTS,
+    JWT_REFRESH_STORE ='keyv', AUTH_USER_STORE, AUTH_USER_STORE_NAME, JWT_REFRESH_STORE_NAME
+  } = options || {});
+
+  ({ setRefreshToken, getRefreshToken, revokeRefreshToken, setRefreshTokenStoreName } = require('./' + JWT_REFRESH_STORE)); // keyv, redis, mongo, knex
+  ({ findUser, updateUser, setAuthUserStoreName } = require('./' + AUTH_USER_STORE)); // mongo, knex
+  
+  if (setRefreshTokenStoreName) setRefreshTokenStoreName(JWT_REFRESH_STORE_NAME)
+  if (setAuthUserStoreName) setAuthUserStoreName(AUTH_USER_STORE_NAME)  
 }
-
-let knex
-if (AUTH_USER_STORE === 'knex') {
-  knex = require('../services/db/knex').get()
-}
-
-const { setToken, getToken, revokeToken } = require('./' + JWT_REFRESH_STORE)
 
 // SameSite=None; must use with Secure;
 // may need to restart browser, TBD set Max-Age, ALTERNATE use res.cookie, Signed?
-const httpOnlyCookie = `HttpOnly;SameSite=${COOKIE_SAMESITE};`
+const httpOnlyCookie = () => `HttpOnly;SameSite=${COOKIE_SAMESITE};`
   + (COOKIE_SECURE ? 'Secure;':'')
   + (COOKIE_MAXAGE ? 'MaxAge='+COOKIE_MAXAGE+';':'')
   + (COOKIE_DOMAIN ? 'domain='+COOKIE_DOMAIN+';':'')
@@ -38,27 +44,7 @@ const httpOnlyCookie = `HttpOnly;SameSite=${COOKIE_SAMESITE};`
 // audience  = 'http://mysoftcorp.in'
 // ip
 
-// We implement stateful refresh_token not stateless ATM
-
-const findUser = async (where) => {
-  if (AUTH_USER_STORE === 'mongo') {
-    if (where.id) where = { _id: new ObjectID(where.id) }
-    return await mongo.db.collection(AUTH_USER_STORE_NAME).findOne(where)
-  } else if (AUTH_USER_STORE === 'knex') {
-    return await knex(AUTH_USER_STORE_NAME).where(where).first()
-  }
-  return null
-}
-
-const updateUser = async (where, payload) => {
-  if (AUTH_USER_STORE === 'mongo') {     
-    if (where.id) where = { _id: new ObjectID(where.id) }
-    return await mongo.db.collection(AUTH_USER_STORE_NAME).updateOne(where, { $set: payload })
-  } else if (AUTH_USER_STORE === 'knex') {
-    return await knex(AUTH_USER_STORE_NAME).where(where).first().update(payload)
-  }
-  return null
-}
+// We implement stateful refresh_token not stateless
 
 // mode: sign, verify
 // type: access, refresh
@@ -104,7 +90,7 @@ const createToken = async (user) => { // Create a tokens & data from user
   options.expiresIn = JWT_REFRESH_EXPIRY
   const refresh_token = jwt.sign({ id }, getSecret('sign', 'refresh'), options) // store only ID in refresh token?
 
-  await setToken(id, refresh_token) // store in DB or Cache
+  await setRefreshToken(id, refresh_token) // store in DB or Cache
   return {
     access_token,
     refresh_token,
@@ -116,8 +102,8 @@ const setTokensToHeader = (res, {access_token, refresh_token}) => {
   const _access_token = `Bearer ${access_token}`
   if (COOKIE_HTTPONLY) {
     res.setHeader('Set-Cookie', [
-      `Authorization=${_access_token};Path=/;`+ httpOnlyCookie,
-      `refresh_token=${refresh_token};Path=${AUTH_REFRESH_URL};`+ httpOnlyCookie // send only if path contains refresh
+      `Authorization=${_access_token};Path=/;`+ httpOnlyCookie(),
+      `refresh_token=${refresh_token};Path=${AUTH_REFRESH_URL};`+ httpOnlyCookie() // send only if path contains refresh
     ])
   } else {
     res.setHeader('Authorization', `${_access_token}`)
@@ -161,7 +147,7 @@ const authRefresh = async (req, res) => { // get refresh token
     const refresh_token = req.cookies?.refresh_token || req.header('refresh_token') || req.query?.refresh_token // check refresh token & user - always stateful          
     const refresh_result = jwt.verify(refresh_token, getSecret('verify', 'refresh'), { algorithm: [JWT_ALG] }) // throw if expired or invalid
     const { id } = refresh_result
-    let refreshToken = await getToken(id)
+    let refreshToken = await getRefreshToken(id)
     if (String(refreshToken) === String(refresh_token)) { // ok... generate new access token & refresh token?
       const user = await findUser({ id })
       const tokens = await createToken(user) // 5 minute expire for login
@@ -190,7 +176,7 @@ const logout = async (req, res) => {
   }
   try {
     if (id) {
-      await revokeToken(id) // clear
+      await revokeRefreshToken(id) // clear
       if (COOKIE_HTTPONLY) {
         res.clearCookie('refresh_token')
         res.clearCookie('Authorization')
@@ -254,7 +240,7 @@ const otp = async (req, res) => { // need to be authentication, body { id: '', p
   return res.status(401).json({ message: 'Error token revoked' })
 }
 
-module.exports = { findUser, updateUser, createToken, setTokensToHeader, revokeToken, authUser, authRefresh, logout, refresh, login, otp, httpOnlyCookie, bcrypt, otplib } // getToken, setToken,
+module.exports = { setupAuth, findUser, updateUser, createToken, setTokensToHeader, authUser, authRefresh, logout, refresh, login, otp, bcrypt, otplib }
 
 // do refresh token check from backend ?
 /*