+++ @es-labs/node/auth/keyv.js
@@ -1,13 +1,6 @@
 'use strict'
 
-const keyv = require('../services/keyv')
-
-// set LRU and Expiry
-
-const setToken = async (id, refresh_token) => await keyv.set(id, refresh_token)
-
-const getToken = async (id) => await keyv.get(id)
-
-const revokeToken = async(id) => await keyv.delete(id)
-
-module.exports = { setToken, getToken, revokeToken }
+const keyv = require('../services/db/keyv') // set LRU and Expiry
+exports.setRefreshToken = async (id, refresh_token) => await keyv.get().set(id, refresh_token)
+exports.getRefreshToken = async (id) => await keyv.get().get(id)
+exports.revokeRefreshToken = async(id) => await keyv.get().delete(id)