+++ @es-labs/node/auth/redis.js
@@ -1,12 +1,6 @@
 'use strict'
-const redis = require('../services/redis')
 
-// set LRU and Expiry
-
-const setToken = async (id, refresh_token) => await redis.set(id, refresh_token)
-
-const getToken = async (id) => await redis.get(id)
-
-const revokeToken = async(id) => await redis.del(id)
-
-module.exports = { setToken, getToken, revokeToken }
+const redis = require('../services/redis') // set LRU and Expiry
+exports.setRefreshToken = async (id, refresh_token) => await redis.get().set(id, refresh_token)
+exports.getRefreshToken = async (id) => await redis.get().get(id)
+exports.revokeRefreshToken = async(id) => await redis.get().del(id)