+++ js-node/expressjs/apps/app-template/tests/integration/category.route.test.js
@@ -13,11 +13,14 @@ beforeAll(async () => {
   await require('@es-labs/node/config')(process.cwd())
   sqldb = await require('@es-labs/node/services/db/knex').open()
   keyv = await require('@es-labs/node/services/db/keyv').open()
+
+  const { createToken, setupAuth } = require('@es-labs/node/auth')
+  setupAuth()
+
   require(APP_PATH + '/common/init')(app, express, global.CONFIG)
   require(APP_PATH + '/common/preRoute')(app, express, global.CONFIG)
   require(APP_PATH + '/router')(app)
 
-  const { createToken } = require('@es-labs/node/auth')
   const tokens = await createToken({ id: 100, groups: 'TestGroup' })
   authObj = {
     Authorization: `Bearer ${tokens.access_token}`,