+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 - [@es-labs/node] allow options to be passed in also, currently uses global.CONFIG through (**@es-labs/node/config**)
   - [@es-labs/node/services/websocket.js](@es-labs/node/services/websocket.js) add onClientConnect function
   - small breaking changes
-    - [@es-labs/node/comms](@es-labs/node/comms) - add setup function, move **../services/webpush.js** here
+    - [@es-labs/node/comms](@es-labs/node/comms) - add setup function
+    - [@es-labs/node/auth](@es-labs/node/auth) - add setup function
+    - [@es-labs/node/services/webpush.js](@es-labs/node/services/webpush.js) - add setup function, move to [@es-labs/node/services/comms/webpush.js](@es-labs/node/services/comms/webpush.js)
     - [@es-labs/node/services/gcp.js](@es-labs/node/services/gcp.js) add setupStorage function
-- [js-node/expressjs/apps/app-template/config/secret/README.md](js-node/expressjs/apps/app-template/config/secret/README.md) - improve oauth configuration
+    - [@es-labs/node/services/redis.js](@es-labs/node/services/redis.js) - add setup function, move to [@es-labs/node/services/db/redis.js](@es-labs/node/services/db/redis.js)
+    - [@es-labs/node/services/keyv.js](@es-labs/node/services/keyv.js) - add setup function, move to [@es-labs/node/services/db/keyv.js](@es-labs/node/services/db/keyv.js)
+    - [js-node/expressjs/apps/app-template/config/secret/README.md](js-node/expressjs/apps/app-template/config/secret/README.md) - improve oauth configuration
 - TBD
   - [js-node/expressjs/router/saml.js](js-node/expressjs/router/saml.js) - improve saml configuration
   - [@es-labs/node/services/keyv.js](@es-labs/node/services/keyv.js) - to improve